requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
Flask-SocketIO==5.5.1
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.15.1
requests==2.32.5
simple-websocket==1.1.0
urllib3==2.6.2
Werkzeug==3.1.4
wsproto==1.3.2
